+++ Classes/Domain/Model/Registration.php
@@ -1,0 +1,30 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OliverKlee\Seminars\Domain\Model;
+
+use TYPO3\CMS\Extbase\Annotation as Extbase;
+use TYPO3\CMS\Extbase\DomainObject\AbstractEntity;
+
+/**
+ * This class represents a registration (or a waiting list entry) for an event.
+ */
+class Registration extends AbstractEntity
+{
+    /**
+     * @var string
+     * @Extbase\Validate("StringLength", options={"maximum": 255})
+     */
+    protected $title = '';
+
+    public function getTitle(): string
+    {
+        return $this->title;
+    }
+
+    public function setTitle(string $name): void
+    {
+        $this->title = $name;
+    }
+}