+++ Tests/Unit/Domain/Model/RegistrationTest.php
@@ -1,0 +1,54 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OliverKlee\Seminars\Tests\Unit\Domain\Model;
+
+use Nimut\TestingFramework\TestCase\UnitTestCase;
+use OliverKlee\Seminars\Domain\Model\Registration;
+use TYPO3\CMS\Extbase\DomainObject\AbstractEntity;
+
+/**
+ * @covers \OliverKlee\Seminars\Domain\Model\Registration
+ */
+final class RegistrationTest extends UnitTestCase
+{
+    /**
+     * @var Registration
+     */
+    private $subject;
+
+    protected function setUp(): void
+    {
+        parent::setUp();
+
+        $this->subject = new Registration();
+    }
+
+    /**
+     * @test
+     */
+    public function isAbstractEntity(): void
+    {
+        self::assertInstanceOf(AbstractEntity::class, $this->subject);
+    }
+
+    /**
+     * @test
+     */
+    public function getTitleInitiallyReturnsEmptyString(): void
+    {
+        self::assertSame('', $this->subject->getTitle());
+    }
+
+    /**
+     * @test
+     */
+    public function setTitleSetsTitle(): void
+    {
+        $value = 'the latest registration';
+        $this->subject->setTitle($value);
+
+        self::assertSame($value, $this->subject->getTitle());
+    }
+}